pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools<81", "setuptools_scm", "build"]

[project]
authors = [{ name = "coreimage", email = "alex@cacko.net" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
description = "image utils"
name = "coreimage"
requires-python = ">=3.11"
dynamic = ["version"]
dependencies = [
    "click>=8.1.3",
    "flake8>=6.0.0",
    "mccabe>=0.7.0",
    "opencv-python>=4.7.0.72",
    "pillow>=9.5.0",
    "semver>=3.0.0",
    "pygments>=2.15.1",
    "questionary>=1.10.0",
    "pyfiglet>=0.8.post1",
    "emoji>=2.4.0",
    "pydantic>=2.7.1",
    "corelog>=0.0.22",
    "corefile>=0.1.18",
    "corestring>=0.3.6",
    "term-image>=0.8.0.1",
    "segno>=1.5.2",
    "spandrel>=0.2",
    "facenet-pytorch>=20.5.2",
    "transformers>=4.41.1",
    "torch>=2.2.2",
    "torchvision>=0.17.2",
    "cytoolz>=1.0.1",
    "olefile>=0.47",
    "six>=1.17",
    "kornia>=0.8.2",
    "timm>=1.0.22",
]
[project.scripts]
coreimage = "coreimage.cli:cli"

[tool.setuptools_scm]
version_file = "version.txt"

[tool.setuptools.package-data]
"*" = ["*.pth"]
